Move the connection timeout test from linux-only to cross-platform as it's now supported on windows
