+++ pkg/network/tracer/tracer_linux_test.go
@@ -2368,8 +2368,7 @@ var failedConnectionsBuildModes = map[ebpftest.BuildMode]struct{}{
 	ebpftest.RuntimeCompiled: {},
 }
 
-func (s *TracerSuite) TestTCPFailureConnectionTimeout() {
-	t := s.T()
+func checkSkipFailureConnectionsTests(t *testing.T) {
 	if _, ok := failedConnectionsBuildModes[ebpftest.GetBuildMode()]; !ok {
 		t.Skip("Skipping test on unsupported build mode: ", ebpftest.GetBuildMode())
 	}
@@ -2377,6 +2376,12 @@ func (s *TracerSuite) TestTCPFailureConnectionTimeout() {
 	if kv < kernel.VersionCode(4, 19, 0) {
 		t.Skip("Skipping test on kernels < 4.19")
 	}
+}
+func (s *TracerSuite) TestTCPFailureConnectionTimeout() {
+	t := s.T()
+
+	checkSkipFailureConnectionsTests(t)
+
 	setupDropTrafficRule(t)
 	cfg := testConfig()
 	cfg.TCPFailedConnectionsEnabled = true
@@ -2398,7 +2403,11 @@ func (s *TracerSuite) TestTCPFailureConnectionTimeout() {
 	t.Cleanup(func() { syscall.Close(sfd) })
 
 	//syscall.TCP_USER_TIMEOUT is 18 but not defined in our linter. Set it to 500ms
-	err = syscall.SetsockoptInt(sfd, syscall.IPPROTO_TCP, 18, 500)
+	if runtime.GOOS == "windows" {
+		err = syscall.SetsockoptInt(sfd, syscall.SOL_SOCKET, 0x1005, 500)
+	} else {
+		err = syscall.SetsockoptInt(sfd, syscall.IPPROTO_TCP, 18, 500)
+	}
 	require.NoError(t, err)
 
 	err = syscall.Connect(sfd, &addr)
@@ -2425,32 +2434,6 @@ func (s *TracerSuite) TestTCPFailureConnectionTimeout() {
 		// 110 is the errno for ETIMEDOUT
 		return findFailedConnection(t, localAddr, srvAddr, conns, 110)
 	}, 3*time.Second, 1000*time.Millisecond, "Failed connection not recorded properly")
-}
-
-func (s *TracerSuite) TestTCPFailureConnectionRefused() {
-	t := s.T()
-	if _, ok := failedConnectionsBuildModes[ebpftest.GetBuildMode()]; !ok {
-		t.Skip("Skipping test on unsupported build mode: ", ebpftest.GetBuildMode())
-	}
-	cfg := testConfig()
-	cfg.TCPFailedConnectionsEnabled = true
-	tr := setupTracer(t, cfg)
-
-	// try to connect to a port where no server is accepting connections
-	srvAddr := "127.0.0.1:9998"
-	conn, err := net.Dial("tcp", srvAddr)
-	if err == nil {
-		conn.Close() // If the connection unexpectedly succeeds, close it immediately.
-		require.Fail(t, "expected connection to be refused, but it succeeded")
-	}
-	require.Error(t, err, "expected connection refused error but got none")
-
-	// Check if the connection was recorded as refused
-	require.Eventually(t, func() bool {
-		conns := getConnections(t, tr)
-		// Check for the refusal record
-		return findFailedConnectionByRemoteAddr(srvAddr, conns, 111)
-	}, 3*time.Second, 100*time.Millisecond, "Failed connection not recorded properly")
 }
 
 func (s *TracerSuite) TestTCPFailureConnectionReset() {
@@ -2497,37 +2480,6 @@ func (s *TracerSuite) TestTCPFailureConnectionReset() {
 	}, 3*time.Second, 100*time.Millisecond, "Failed connection not recorded properly")
 
 	require.NoError(t, c.Close(), "error closing client connection")
-}
-
-// findFailedConnection is a utility function to find a failed connection based on specific TCP error codes
-func findFailedConnection(t *testing.T, local, remote string, conns *network.Connections, errorCode uint32) bool {
-	// Extract the address and port from the net.Addr types
-	localAddrPort, err := netip.ParseAddrPort(local)
-	if err != nil {
-		t.Logf("Failed to parse local address: %v", err)
-		return false
-	}
-	remoteAddrPort, err := netip.ParseAddrPort(remote)
-	if err != nil {
-		t.Logf("Failed to parse remote address: %v", err)
-		return false
-	}
-
-	failureFilter := func(cs network.ConnectionStats) bool {
-		localMatch := netip.AddrPortFrom(cs.Source.Addr, cs.SPort) == localAddrPort
-		remoteMatch := netip.AddrPortFrom(cs.Dest.Addr, cs.DPort) == remoteAddrPort
-		return localMatch && remoteMatch && cs.TCPFailures[errorCode] > 0
-	}
-
-	return network.FirstConnection(conns, failureFilter) != nil
-}
-
-// for some failed connections we don't know the local addr/port so we need to search by remote addr only
-func findFailedConnectionByRemoteAddr(remoteAddr string, conns *network.Connections, errorCode uint32) bool {
-	failureFilter := func(cs network.ConnectionStats) bool {
-		return netip.MustParseAddrPort(remoteAddr) == netip.AddrPortFrom(cs.Dest.Addr, cs.DPort) && cs.TCPFailures[errorCode] > 0
-	}
-	return network.FirstConnection(conns, failureFilter) != nil
 }
 
 func setupDropTrafficRule(tb testing.TB) (ns string) {
