+++ pkg/network/tracer/tracer_windows_test.go
@@ -8,6 +8,9 @@
 package tracer
 
 import (
+	"github.com/DataDog/datadog-agent/pkg/network/testutil"
+	"testing"
+
 	sysconfigtypes "github.com/DataDog/datadog-agent/cmd/system-probe/config/types"
 	"github.com/DataDog/datadog-agent/pkg/network/config"
 	"github.com/DataDog/datadog-agent/pkg/network/driver"
@@ -29,4 +32,25 @@ func classificationSupported(config *config.Config) bool {
 func testConfig() *config.Config {
 	cfg := config.New()
 	return cfg
+}
+
+func setupDropTrafficRule(tb testing.TB) (ns string) {
+	//
+	// note.  This does not seem to function as advertised; localhost traffic is not being
+	// blocked.  More re
+	tb.Cleanup(func() {
+		cmds := []string{
+			"powershell -c \"Remove-NetFirewallRule -DisplayName 'Datadog Test Rule'\"",
+		}
+		testutil.RunCommands(tb, cmds, false)
+	})
+	cmds := []string{
+		"powershell -c \"New-NetFirewallRule -DisplayName 'Datadog Test Rule' -Direction Outbound -Action Block -Profile Any -RemotePort 10000 -Protocol TCP\"",
+	}
+	testutil.RunCommands(tb, cmds, false)
+	return
+}
+
+func checkSkipFailureConnectionsTests(t *testing.T) {
+	return
 }
